+++ backend/src/endpoints/role/router.py
@@ -5,9 +5,10 @@ from fastapi import APIRouter, Body
 from fastapi.responses import JSONResponse
 
 from src.endpoints.auth.depends import current_user
+from src.endpoints.role.permissions import role_permission
 
 from .depends import role_service_factory
-from .models import RoleModel, RolePermissionsEnum
+from .models import RolePermissionsEnum
 from .schemas import CreateRoleSchema, ReadRoleSchema, UpdateRoleSchema
 
 router = APIRouter(prefix="/roles_group", tags=["role"])
@@ -20,19 +21,27 @@ async def get_role(role_id: UUID, role_service: role_service_factory) -> ReadRol
     return ReadRoleSchema.model_validate(await role_service.get_one_by_field(role_id))
 
 
+@router.get("/permissions/me/{group_id}")
+async def get_my_permissions_in_group(
+    role_service: role_service_factory, user: current_user, group_id: UUID
+):
+    await role_service.get_user_permissions_in_group(user.id, group_id)
+
+
 @router.post("/{group_id}")
 async def create_role(
     group_id: UUID,
     user: current_user,
     role: Annotated[CreateRoleSchema, Body()],
     role_service: role_service_factory,
+    permission: role_permission,
 ) -> ReadRoleSchema:
-    await role_service.permission.check_user_has_permission(
-        RolePermissionsEnum.CAN_EDIT_ROLES.value, user.id, group_id
+    await permission.check_user_has_permission(
+        RolePermissionsEnum.CAN_EDIT_ROLES, user.id, group_id
     )
 
-    role_db = RoleModel(**role.model_dump(), group_id=group_id)
-    new_role = await role_service.create(role_db)
+    role.group_id = group_id
+    new_role = await role_service.create(role)
 
     return ReadRoleSchema.model_validate(new_role)
 
@@ -44,12 +53,18 @@ async def update_role(
     user: current_user,
     role: Annotated[UpdateRoleSchema, Body()],
     role_service: role_service_factory,
+    permission: role_permission,
 ):
-    await role_service.permission.check_user_has_permission(
-        RolePermissionsEnum.CAN_EDIT_ROLES.value, user.id, group_id
+    await permission.check_user_has_permission(
+        RolePermissionsEnum.CAN_EDIT_ROLES, user.id, group_id
     )
 
-    await role_service.update_one_by_id(role.model_dump(exclude_none=True), id=role_id)
+    await role_service.update(
+        UpdateRoleSchema(
+            id=role_id,
+            **role.model_dump(exclude_none=True),
+        )
+    )
 
     return JSONResponse(status_code=200, content={"message": "Group updated"})
 
@@ -58,7 +73,12 @@ async def update_role(
 async def delete_role(
     role_id: UUID,
     role_service: role_service_factory,
+    permission: role_permission,
 ):
+    await permission.check_user_has_permission(
+        RolePermissionsEnum.CAN_EDIT_ROLES, user.id, group_id
+    )
+
     await role_service.delete(role_id)
 
     return JSONResponse(status_code=200, content={"message": "Group deleted"})
